[#11] Improve Redis object relationship handling
This commit improves the Redis object handling in the Familia::Horreum
module by:

- Splitting Redis-related functionality into separate modules
- Removing deprecated Ruby syntax and using modern alternatives
- Updating method implementations for better performance and readability
- Enhancing code organization and reducing duplication
